backend/requirements.txt
Django==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.3
